Refactor: move snapshot handling from Engine to SnapshotHandler
